+++ cli/srecli/group/metrics.py
@@ -1,6 +1,14 @@
 import click
 
-from srecli.group import ContextObj, ViewCommand, cli_response
+from srecli.group import (
+    ContextObj, 
+    ViewCommand, 
+    cli_response, 
+)
+from srecli.service.constants import (
+    ServiceOperationType, 
+    OPERATION_STATUS_HINT,
+)
 from srecli.service.adapter_client import SREResponse
 
 
@@ -14,7 +22,9 @@ def metrics():
     name='update',
 )
 @cli_response(
-    hint="Use 'sre service_operation status --operation metrics_update' to check execution status",
+    hint=OPERATION_STATUS_HINT.format(
+        operation_type=ServiceOperationType.UPDATE_METRICS.value[0],
+    ),
 )
 def update(
     ctx: ContextObj,